requirements.txt
pandas
numpy
pyyaml
langchain
langchain-core
langchain-openai
openai
python-dotenv
pydantic
fastapi
uvicorn[standard]
